+++ packages/dso-toolkit/src/components/logo/logo.models.ts
@@ -1,0 +1,4 @@
+export interface Logo {
+  label?: string;
+  ribbon?: string;
+}